Sécurisation plugin delete table (#9)
* Sécurisation plugin delete table

* Secure delete table clean

* nonce in action duplicate-table

* nonce in action delete webshop

* nonce in action  delete link

* nonce in action edit table

* nonce in action edit webshop

* nonce in action edit link

---------

+++ site/wp-content/plugins/affieasy/views/admin/list-table.php
@@ -22,16 +22,19 @@ wp_localize_script( 'list-table-script', 'translations', array(
 
 wp_enqueue_style('wp-jquery-ui-dialog');
 
-$id = isset($_GET['id']) ? sanitize_key($_GET['id']) : null;
-$action = isset($_GET['action']) ?  sanitize_key($_GET['action']) : null;
+$id = isset($_REQUEST['id']) ? sanitize_key($_REQUEST['id']) : null;
+$action = isset($_REQUEST['action']) ?  sanitize_key($_REQUEST['action']) : null;
 
 $isValidDeleteAction = $action === 'delete-table' && is_numeric($id);
 
 if (is_numeric($id)) {
     $dbManager = new AFES_DbManager();
-    if ($action === 'delete-table') {
+    
+    $nonce = isset($_REQUEST['_wpnonce']) ? $_REQUEST['_wpnonce'] : null;
+
+    if ($action === 'delete-table' && wp_verify_nonce( $nonce, 'my-nonce') ) {
         $dbManager->delete_table($id);
-    } else if ($action === 'duplicate-table') {
+    } else if ($action === 'duplicate-table' && wp_verify_nonce( $nonce, 'my-nonce')) {
         $dbManager->duplicate_table($id);
     }
 }
@@ -69,3 +72,15 @@ $tableList = new AFES_TableList();
         ?>
     </form>
 </div>
+<script>
+    jQuery(($) => {
+
+        $('.delete-table-confirm').click(function(e){
+        
+            if (!confirm('<?php esc_html_e('Are you sure you want to delete the table?', 'affieasy'); ?>')){
+                e.preventDefault();
+            }    
+
+        });
+    });
+</script>
